backend/requirements.txt
# FPL Assistant Backend Requirements - Complete for Python 3.13

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Database
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
alembic==1.13.3

# HTTP Requests
requests==2.32.3
aiohttp==3.10.10
httpx==0.27.2

# Data Processing (Python 3.13 compatible versions)
pandas==2.2.3
numpy==2.1.3

# Scheduling
APScheduler==3.11.0

# Caching
redis==5.2.0

# Environment Variables
python-dotenv==1.0.1

# Scraping
beautifulsoup4==4.12.3
lxml==5.3.0

# Logging
loguru==0.7.3

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# Data Validation
pydantic==2.10.0

# Date/Time utilities
python-dateutil==2.9.0

# Math/Stats (Python 3.13 compatible)
scipy==1.14.1

# Development
black==24.10.0
flake8==7.1.1